// src/app/models/browser.ts
export class BrowserData {

    browserAcceptHeader: string;
    browserIP: string;
    browserJavaEnabled: boolean;
    browserLanguage: string;
    browserColorDepth: number;
    browserScreenHeight: number;
    browserScreenWidth: number;
    browserTZ: string;
    browserUserAgent: string;
  }
  